+++ virtualizarr/manifests/store.py
@@ -138,6 +138,11 @@ class ManifestStore(Store):
         return f"ManifestStore(group={self._group}, registry={self._registry})"
 
     @property
+    def group(self) -> ManifestGroup:
+        """The root [ManifestGroup][virtualizarr.manifests.ManifestGroup] this store serves."""
+        return self._group
+
+    @property
     def nbytes_virtual(self) -> int:
         """Size required to hold these references in memory in bytes. See [ManifestGroup.nbytes_virtual][virtualizarr.manifests.ManifestGroup.nbytes_virtual]."""
         return self._group.nbytes_virtual